+++ pytket/pytket/passes/__init__.py
@@ -18,3 +18,4 @@ from pytket._tket.passes import *
 
 from .script import compilation_pass_from_script, compilation_pass_grammar
 from .auto_rebase import auto_rebase_pass, auto_squash_pass
+from .superpass import superpass